a few 64-bit compatibility fixes for vs2005
git-svn-id: file://localhost/home/tolik/svnrepo/trunk/opencv@587 c5418bc3-7546-4aae-8bad-fa6f2f3e53e6

+++ src/cvaux/vs/blobtrackanalysisior.cpp
@@ -106,7 +106,7 @@ public:
                 strcat(m_Desc,": ");
                 strcat(m_Desc,str);
                 strcat(m_Desc,"\n");
-                rest = MAX_DESC - strlen(m_Desc) - 1;
+                rest = MAX_DESC - (int)strlen(m_Desc) - 1;
             }
         }/* next analizer */
         if(m_Desc[0]!=0)return m_Desc;
